fix: refactor directory
