+++ frontend/src/MyPage/MyPage.tsx
@@ -38,7 +38,7 @@ const MyPage = () => {
         <div className={styles.right}>
             <div className={styles.right__inner}>
                 {/* 클릭된 버튼의 name으로 buttonList에서 component를 찾아 Render 
-                    추후 항목이 추가되면 buttonList를 수정하면 됩니다.*/}
+                    추후 항목이 추가되면 buttonList를 수정하면 됨.*/}
                 {buttonList.filter(button => button.name == buttonClickState)[0].component}
             </div>
         </div>
